+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added input_file parameter to Console.input
+- Added password parameter to Console.input
 - Added description parameter to Progress.update
+- Added rich.prompt
 
 ## [3.3.2] - 2020-07-14